requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
pytest
pytest-django
python-dotenv
pydantic_settings